[DOP-19907] - remove markers
