+++ .env.docker
@@ -28,6 +28,13 @@ TEST_S3_ACCESS_KEY=syncmaster
 TEST_S3_SECRET_KEY=changeme
 TEST_S3_BUCKET=syncmaster
 
+# needed for test-postgres container
+POSTGRES_HOST=test-postgres
+POSTGRES_PORT=5432
+POSTGRES_DB=syncmaster
+POSTGRES_USER=syncmaster
+POSTGRES_PASSWORD=changeme
+
 TEST_POSTGRES_HOST_FOR_CONFTEST=test-postgres
 TEST_POSTGRES_PORT_FOR_CONFTEST=5432
 TEST_POSTGRES_HOST_FOR_WORKER=test-postgres
