+++ .env.local
@@ -1,11 +1,14 @@
 export TZ=UTC
 export ENV=LOCAL
 
+# Debug
 export SYNCMASTER__SERVER__DEBUG=true
 
+# Logging
 export SYNCMASTER__SERVER__LOGGING__SETUP=True
 export SYNCMASTER__SERVER__LOGGING__PRESET=json
 
+# Postgres
 export SYNCMASTER__DATABASE__URL=postgresql+asyncpg://syncmaster:changeme@localhost:5432/syncmaster
 
 # RabbitMQ
