--- syncmaster/backend/api/deps.py
@@ -1,14 +1,0 @@
-# SPDX-FileCopyrightText: 2023-2024 MTS PJSC
-# SPDX-License-Identifier: Apache-2.0
-
-
-class DatabaseEngineMarker:
-    pass
-
-
-class DatabaseSessionMarker:
-    pass
-
-
-class UnitOfWorkMarker:
-    pass
