+++ syncmaster/backend/api/v1/queue.py
@@ -2,7 +2,6 @@
 # SPDX-License-Identifier: Apache-2.0
 from fastapi import APIRouter, Depends, Query, status
 
-from syncmaster.backend.api.deps import UnitOfWorkMarker
 from syncmaster.backend.services import UnitOfWork, get_user
 from syncmaster.db.models import User
 from syncmaster.db.utils import Permission
@@ -27,7 +26,7 @@ async def read_queues(
     page: int = Query(gt=0, default=1),
     page_size: int = Query(gt=0, le=200, default=20),
     current_user: User = Depends(get_user(is_active=True)),
-    unit_of_work: UnitOfWork = Depends(UnitOfWorkMarker),
+    unit_of_work: UnitOfWork = Depends(UnitOfWork),
     search_query: str | None = Query(
         None,
         title="Search Query",
@@ -55,7 +54,7 @@ async def read_queues(
 async def read_queue(
     queue_id: int,
     current_user: User = Depends(get_user(is_active=True)),
-    unit_of_work: UnitOfWork = Depends(UnitOfWorkMarker),
+    unit_of_work: UnitOfWork = Depends(UnitOfWork),
 ) -> ReadQueueSchema:
     resource_role = await unit_of_work.queue.get_resource_permission(
         user=current_user,
@@ -75,7 +74,7 @@ async def read_queue(
 async def create_queue(
     queue_data: CreateQueueSchema,
     current_user: User = Depends(get_user(is_active=True)),
-    unit_of_work: UnitOfWork = Depends(UnitOfWorkMarker),
+    unit_of_work: UnitOfWork = Depends(UnitOfWork),
 ) -> ReadQueueSchema:
     group_permission = await unit_of_work.queue.get_group_permission(
         user=current_user,
@@ -98,7 +97,7 @@ async def update_queue(
     queue_id: int,
     queue_data: UpdateQueueSchema,
     current_user: User = Depends(get_user(is_active=True)),
-    unit_of_work: UnitOfWork = Depends(UnitOfWorkMarker),
+    unit_of_work: UnitOfWork = Depends(UnitOfWork),
 ) -> ReadQueueSchema:
     resource_role = await unit_of_work.queue.get_resource_permission(
         user=current_user,
@@ -123,7 +122,7 @@ async def update_queue(
 async def delete_queue(
     queue_id: int,
     current_user: User = Depends(get_user(is_active=True)),
-    unit_of_work: UnitOfWork = Depends(UnitOfWorkMarker),
+    unit_of_work: UnitOfWork = Depends(UnitOfWork),
 ) -> StatusResponseSchema:
     resource_role = await unit_of_work.queue.get_resource_permission(
         user=current_user,
