+++ syncmaster/backend/api/v1/runs.py
@@ -7,7 +7,6 @@ from asgi_correlation_id import correlation_id
 from fastapi import APIRouter, Depends, Query
 from kombu.exceptions import KombuError
 
-from syncmaster.backend.api.deps import UnitOfWorkMarker
 from syncmaster.backend.services import UnitOfWork, get_user
 from syncmaster.db.models import Status, User
 from syncmaster.db.utils import Permission
@@ -21,7 +20,6 @@ from syncmaster.schemas.v1.transfers.run import (
     ReadRunSchema,
     RunPageSchema,
 )
-from syncmaster.utils import render_log_url
 from syncmaster.worker.config import celery
 
 router = APIRouter(tags=["Runs"], responses=get_error_responses())
@@ -30,13 +28,13 @@ router = APIRouter(tags=["Runs"], responses=get_error_responses())
 @router.get("/runs")
 async def read_runs(
     transfer_id: int,
+    unit_of_work: UnitOfWork = Depends(UnitOfWork),
     page: int = Query(gt=0, default=1),
     page_size: int = Query(gt=0, le=200, default=20),
     status: list[Status] | None = Query(default=None),
     started_at_since: datetime | None = Query(default=None),
     started_at_until: datetime | None = Query(default=None),
     current_user: User = Depends(get_user(is_active=True)),
-    unit_of_work: UnitOfWork = Depends(UnitOfWorkMarker),
 ) -> RunPageSchema:
     """Return runs of transfer with pagination"""
     resource_rule = await unit_of_work.transfer.get_resource_permission(
@@ -62,8 +60,8 @@ async def read_runs(
 @router.get("/runs/{run_id}")
 async def read_run(
     run_id: int,
+    unit_of_work: UnitOfWork = Depends(UnitOfWork),
     current_user: User = Depends(get_user(is_active=True)),
-    unit_of_work: UnitOfWork = Depends(UnitOfWorkMarker),
 ) -> ReadRunSchema:
     run = await unit_of_work.run.read_by_id(run_id=run_id)
 
@@ -81,8 +79,8 @@ async def read_run(
 @router.post("/runs")
 async def start_run(
     create_run_data: CreateRunSchema,
+    unit_of_work: UnitOfWork = Depends(UnitOfWork),
     current_user: User = Depends(get_user(is_active=True)),
-    unit_of_work: UnitOfWork = Depends(UnitOfWorkMarker),
 ) -> ReadRunSchema:
     # Check: user can start transfer
     resource_rule = await unit_of_work.transfer.get_resource_permission(
@@ -119,11 +117,6 @@ async def start_run(
             source_creds=ReadAuthDataSchema(auth_data=credentials_source).dict(),
             target_creds=ReadAuthDataSchema(auth_data=credentials_target).dict(),
         )
-        log_url = render_log_url(run=run, correlation_id=request_correlation_id)
-        run = await unit_of_work.run.update(
-            run_id=run.id,
-            log_url=log_url,
-        )
 
     try:
         celery.send_task(
@@ -144,8 +137,8 @@ async def start_run(
 @router.post("/runs/{run_id}/stop")
 async def stop_run(
     run_id: int,
+    unit_of_work: UnitOfWork = Depends(UnitOfWork),
     current_user: User = Depends(get_user(is_active=True)),
-    unit_of_work: UnitOfWork = Depends(UnitOfWorkMarker),
 ) -> ReadRunSchema:
     run = await unit_of_work.run.read_by_id(run_id=run_id)
