+++ syncmaster/backend/dependencies/__init__.py
@@ -1,0 +1,4 @@
+# SPDX-FileCopyrightText: 2023-2024 MTS PJSC
+# SPDX-License-Identifier: Apache-2.0
+
+from syncmaster.backend.dependencies.stub import Stub
