+++ syncmaster/db/base.py
