--- syncmaster/db/mixins/timestamp.py
@@ -1,17 +1,0 @@
-# SPDX-FileCopyrightText: 2023-2024 MTS PJSC
-# SPDX-License-Identifier: Apache-2.0
-from datetime import datetime
-
-from sqlalchemy import DateTime, func
-from sqlalchemy.orm import Mapped, mapped_column
-
-
-class TimestampMixin:
-    created_at: Mapped[datetime] = mapped_column(DateTime, server_default=func.now(), nullable=False)
-    updated_at: Mapped[datetime] = mapped_column(
-        DateTime,
-        server_default=func.now(),
-        server_onupdate=func.now(),
-        onupdate=datetime.now,
-        nullable=False,
-    )
