+++ tests/test_unit/test_runs/test_create_run.py
@@ -126,12 +126,6 @@ async def test_superuser_can_create_run(
     mocker,
     monkeypatch,
 ) -> None:
-    # Arrange
-    monkeypatch.setenv(
-        "LOG_URL_TEMPLATE",
-        "https://grafana.example.com?correlation_id={{ correlation_id }}&run_id={{ run.id }}",
-    )
-
     mocker.patch("syncmaster.worker.config.celery.send_task")
 
     # Act
@@ -157,8 +151,6 @@ async def test_superuser_can_create_run(
         "transfer_dump": run.transfer_dump,
     }
     assert result.status_code == 200
-    assert "correlation_id" in result.json().get("log_url")
-    assert "run_id" in result.json().get("log_url")
 
 
 async def test_unauthorized_user_cannot_create_run(
